add numutils for pow2 and base 2 or any base conversions
